requirements.txt
langchain==0.3.13
langchain-openai==0.2.10
langchain-pinecone==0.2.0
langchain-community==0.3.4
pinecone-client==5.0.1
openai==1.55.3
pandas==2.2.3
pydantic==2.10.2
python-dotenv==1.0.1
streamlit==1.40.2
transformers==4.46.3
semantic-router==0.0.72
openpyxl==3.1.3
beautifulsoup4==4.12.3
pymupdf==1.24.14